BaseRepository added
